+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [0.0.16] - 2024-11-22
+
+- Fix bug where mutable not being set explicitly raises
+- Also, make it so models don't default to mutable false when processed by a model type
+
 ## [0.0.15] - 2024-11-20
 
 - Move entity attributes type declaration helpers from Command::EntityHelpers to Entity for convenience.